% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mass_isotope.R
\name{adduct_shift}
\alias{adduct_shift}
\title{Mass shift for a single cation adduct}
\usage{
adduct_shift(adduct = c("Na", "K", "NH4"))
}
\arguments{
\item{adduct}{One of `"Na"`, `"K"`, `"NH4"`.}
}
\value{
The mass shift, in Da.
}
\description{
Negative-mode adduct where one cation replaces one backbone hydrogen:
`+Na` +21.981944 Da, `+K` +37.955882 Da, `+NH4` +17.026549 Da.
}
\seealso{
[match_ms1()]
}
