% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_standard.R
\name{build_is_metabolites}
\alias{build_is_metabolites}
\title{Library entries for internal standards}
\usage{
build_is_metabolites(is_records, dict = STANDARD_DICT)
}
\arguments{
\item{is_records}{\code{parse_sequence_set()$internal_standards}.}

\item{dict}{Chemistry dictionary.}
}
\value{
A list of metabolite objects with ids \code{IS01}, \code{IS02}, ...,
\code{kind = "internal_standard"}, no truncations or fragments.
}
\description{
Library entries for internal standards
}
