% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oligo_io.R
\name{format_triplet}
\alias{format_triplet}
\title{Render an oligo_spec back to triplet notation}
\usage{
format_triplet(spec)
}
\arguments{
\item{spec}{An `oligo_spec` list (see [parse_input()]).}
}
\value{
A single triplet-notation string, e.g. `"Ge-uAn-sGn-..."`.
}
\description{
Render an oligo_spec back to triplet notation
}
\seealso{
[parse_input()] (`notation = "triplet"`), [format_spec()]
}
