% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabolites.R
\name{metabolite_table}
\alias{metabolite_table}
\title{Flatten a metabolite list to a display data.frame}
\usage{
metabolite_table(mets)
}
\arguments{
\item{mets}{A list of metabolite objects (see [generate_metabolites()]).}
}
\value{
A data.frame, one row per metabolite: `id`, `name`, `kind`,
  `n`, `n_ps`, `n_po`, `bases` (collapsed to a single string),
  `modification`, `site`.
}
\description{
Flatten a metabolite list to a display data.frame
}
\seealso{
[generate_metabolites()]
}
