% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mass_isotope.R
\name{ps_oxid_formula}
\alias{ps_oxid_formula}
\title{Molecular formula after k phosphorothioate-to-phosphodiester oxidations}
\usage{
ps_oxid_formula(formula_vec, k)
}
\arguments{
\item{formula_vec}{A full element formula vector (see
\code{\link[=metabolite_mass_info]{metabolite_mass_info()}}).}

\item{k}{Number of desulfurization events (subtracted from S, added
to O).}
}
\value{
The formula vector after \code{k} oxidations.
}
\description{
Molecular formula after k phosphorothioate-to-phosphodiester oxidations
}
\seealso{
\code{\link[=ps_oxid_mass]{ps_oxid_mass()}}, \code{\link[=ps_oxidation_series]{ps_oxidation_series()}}
}
