% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemistry_dict.R
\name{scale_formula}
\alias{scale_formula}
\title{Scale a molecular formula by an integer}
\usage{
scale_formula(a, k)
}
\arguments{
\item{a}{A formula vector (compact or full).}

\item{k}{Scale factor, e.g. the number of repeats of a residue.}
}
\value{
\code{a * k} as a full element vector.
}
\description{
Scale a molecular formula by an integer
}
