% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabolites.R
\name{truncate_3p}
\alias{truncate_3p}
\title{Truncate an oligo_spec from the 3' end (exonuclease truncation)}
\usage{
truncate_3p(spec, k)
}
\arguments{
\item{spec}{An \code{oligo_spec} list (see \code{\link[=parse_input]{parse_input()}}).}

\item{k}{Number of nucleotides to remove, \code{1} to \code{n-1}.}
}
\value{
A truncated spec-like list (\code{bases}, \code{sugars}, \code{linkages},
\code{conj5}, \code{conj3 = "none"} (lost with the 3' end), \code{n}).
}
\description{
Truncate an oligo_spec from the 3' end (exonuclease truncation)
}
\seealso{
\code{\link[=truncate_5p]{truncate_5p()}}, \code{\link[=generate_metabolites]{generate_metabolites()}}
}
