% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabolites.R
\name{truncate_5p}
\alias{truncate_5p}
\title{Truncate an oligo_spec from the 5' end (exonuclease truncation)}
\usage{
truncate_5p(spec, k)
}
\arguments{
\item{spec}{An `oligo_spec` list (see [parse_input()]).}

\item{k}{Number of nucleotides to remove, `1` to `n-1`.}
}
\value{
A truncated spec-like list (`bases`, `sugars`, `linkages`,
  `conj5 = "none"` (lost with the 5' end), `conj3`, `n`).
}
\description{
Truncate an oligo_spec from the 5' end (exonuclease truncation)
}
\seealso{
[truncate_3p()], [generate_metabolites()]
}
