% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemistry_dict.R
\name{add_formulas}
\alias{add_formulas}
\title{Add two molecular formulas}
\usage{
add_formulas(a, b)
}
\arguments{
\item{a, b}{Formula vectors (compact or full).}
}
\value{
`a + b` as a full element vector.
}
\description{
Add two molecular formulas
}
