% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mass_isotope.R
\name{ps_oxid_mass}
\alias{ps_oxid_mass}
\title{Mass after k phosphorothioate-to-phosphodiester desulfurizations}
\usage{
ps_oxid_mass(mono_mass, k)
}
\arguments{
\item{mono_mass}{Monoisotopic mass before oxidation, in Da.}

\item{k}{Number of desulfurization events.}
}
\value{
\code{mono_mass - k * 15.977157}.
}
\description{
Each PS -> PO oxidation event replaces one sulfur with one oxygen
(-15.977157 Da).
}
\seealso{
\code{\link[=ps_oxid_formula]{ps_oxid_formula()}}, \code{\link[=ps_oxidation_series]{ps_oxidation_series()}}
}
