% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabolites.R
\name{truncate_3p}
\alias{truncate_3p}
\title{Truncate an oligo_spec from the 3' end (exonuclease truncation)}
\usage{
truncate_3p(spec, k)
}
\arguments{
\item{spec}{An `oligo_spec` list (see [parse_input()]).}

\item{k}{Number of nucleotides to remove, `1` to `n-1`.}
}
\value{
A truncated spec-like list (`bases`, `sugars`, `linkages`,
  `conj5`, `conj3 = "none"` (lost with the 3' end), `n`).
}
\description{
Truncate an oligo_spec from the 3' end (exonuclease truncation)
}
\seealso{
[truncate_5p()], [generate_metabolites()]
}
