% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabolites.R
\name{truncate_5p}
\alias{truncate_5p}
\title{Truncate an oligo_spec from the 5' end (exonuclease truncation)}
\usage{
truncate_5p(spec, k)
}
\arguments{
\item{spec}{An \code{oligo_spec} list (see \code{\link[=parse_input]{parse_input()}}).}

\item{k}{Number of nucleotides to remove, \code{1} to \code{n-1}.}
}
\value{
A truncated spec-like list (\code{bases}, \code{sugars}, \code{linkages},
\code{conj5 = "none"} (lost with the 5' end), \code{conj3}, \code{n}).
}
\description{
Truncate an oligo_spec from the 5' end (exonuclease truncation)
}
\seealso{
\code{\link[=truncate_3p]{truncate_3p()}}, \code{\link[=generate_metabolites]{generate_metabolites()}}
}
